asl_extended/requirements.txt
tensorflow
opencv-python<4.12.0
numpy<2.0
scikit-learn
Pillow
matplotlib
jupyter
fiftyone
mediapipe<0.11.0

